# refactor: simplify database test infrastructure (Phase 2.4)