--- a/package.json
+++ b/package.json
@@ -174,6 +174,7 @@
     "@playwright/test": "^1.54.2",
     "@vitest/coverage-v8": "^3.2.4",
     "@vitest/ui": "^3.2.4",
+    "better-sqlite3": "^12.2.0",
     "cross-env": "^7.0.3",
     "eslint": "^9.15.0",
     "express": "^5.1.0",