@damilola-elegbede 2025-08-13
@coderabbitai resolve

## Addressing CodeRabbit Review Comments

Found 4 actionable items from CodeRabbit review. Now addressing:

### Code Quality (3 items)
- Fix hard-coded '@tests/helpers/db' import paths - compute relative paths instead
- Preserve original variable names when replacing factory pattern usage
- Add error handling for schema file loading

### Data Integrity (1 item)
- Fix foreign key inconsistency in access_tokens fixture data

Will apply these fixes in organized commits shortly.